+++ R/openmx_functions.R
@@ -252,7 +252,7 @@ as_ram.data.frame <- function(x, groups = NULL, data = NULL, ...){
 }
 
 
-mx_thresholds <- function(df, ...){
+.mx_thresholds_internal <- function(df, ...){
   ord_vars <- sapply(df, inherits, what = "ordered")
   ord_var_nam <- names(df)[ord_vars]
   lapply(ord_var_nam, function(nam){